blog_bot/requirements.txt
anthropic>=0.28.0
requests>=2.31.0
cloudinary>=1.40.0
python-dotenv>=1.0.0
